Break up RGB standards into smaller modules
